refactor(debug_server): extract OTA endpoint family (refs #332, Wave 23b)
Second per-family extract from debug_server.c, stacked on the K144
extract (PR #336).  Same shape: handlers + URI registration moved
to debug_server_ota.{c,h}; debug_server.c shrinks accordingly.

  GET  /ota/check  — query Dragon for available firmware update
  POST /ota/apply  — schedule the update (reboots into fresh DMA heap)

debug_server.c:  4,293 → 4,202 LOC (−91)
debug_server_ota.c:  130 LOC (new)

Mechanical move, identical behavior.  Only call-site change is
`check_auth(req)` → `tab5_debug_check_auth(req)` inside the moved file.

## Verification on hardware

build: clean — only debug_server_ota.c compiled, debug_server.c relinked.
flash: OK to /dev/ttyACM0, boot uptime 18 s.

GET /ota/check  → {current:"0.8.0", partition:"ota_0",
                    error:"ESP_ERR_NOT_SUPPORTED"}  ← Dragon doesn't
                                                     have OTA configured
                                                     for this dev unit;
                                                     route resolves +
                                                     handler returns
                                                     identical shape.

GET /m5         → still works (K144 family from PR #336 also live).

story_smoke (14 steps): 14/14 PASS in 99 s.

## Cumulative on this branch (stacked on #336)

K144 (4 endpoints) + OTA (2 endpoints) extracted.  6 of 72 endpoints
now live in family files; debug_server.c down ~325 LOC from origin/main
baseline (4520 → 4202).  5 families remaining (camera, video/call,
voice/chat/mode, settings/nvs, codec).
